Augment node-brigadier's StringReader with readLocationLiteral method
